Refactored c ffi
